requirements.txt
customtkinter
requests
beautifulsoup4
yt-dlp
ffmpeg-python

